[refactor] Rename login page header folder
