+++ src/containers/Spotlight/Header/Routes.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Router, Route, Switch } from 'react-router-dom';
 import history from 'utils/history';
-import Login from 'containers/Spotlight/Header/Login';
+import LoginPage from 'containers/Spotlight/Header/LoginPage';
 import ExplorePage from 'containers/Spotlight/Header/ExplorePage';
 import PlanningPage from 'containers/Spotlight/Header/PlanningPage';
 import TravelWallPage from 'containers/Spotlight/Header/TravelWallPage';
@@ -14,7 +14,7 @@ import { routePathConfig } from 'containers/Spotlight/Content/Routes';
 export default () => (
   <Router history={history}>
     <Switch>
-      <Route exact path={routePathConfig.login} component={Login} />
+      <Route exact path={routePathConfig.login} component={LoginPage} />
       <Route exact path={routePathConfig.explorePagePath} component={ExplorePage} />
       <Route exact path={routePathConfig.planningPagePath} component={PlanningPage} />
       <Route exact path={routePathConfig.travelWallPagePath} component={TravelWallPage} />
